Make WordIterator and Snippet::MatchPositions use size_t instead of int for offsets into strings.  This avoids some casts.  I also added a typedef for Snippet::MatchPosition which cleans up a bit of the calling code a little. Review URL: http://codereview.chromium.org/13064
git-svn-id: svn://svn.chromium.org/chrome/trunk/src@6260 0039d316-1c4b-4281-b951-d872f2087c98

+++ chrome/browser/history_view.cc
@@ -123,8 +123,8 @@ class SnippetRenderer : public views::View {
                  int x,
                  int y,
                  Snippet::MatchPositions::const_iterator match_iter,
-                 int start,
-                 int end);
+                 size_t start,
+                 size_t end);
 
   DISALLOW_EVIL_CONSTRUCTORS(SnippetRenderer);
 };
@@ -180,9 +180,12 @@ void SnippetRenderer::Paint(ChromeCanvas* canvas) {
 }
 
 int SnippetRenderer::ProcessRun(
-    ChromeCanvas* canvas, int x, int y,
+    ChromeCanvas* canvas,
+    int x,
+    int y,
     Snippet::MatchPositions::const_iterator match_iter,
-    int start, int end) {
+    size_t start,
+    size_t end) {
   int total_width = 0;
 
   while (start < end) {
@@ -196,7 +199,7 @@ int SnippetRenderer::ProcessRun(
     // Determine the next substring to process by examining whether
     // we're before a match or within a match.
     ChromeFont* font = &text_font_;
-    int next = end;
+    size_t next = end;
     if (match_iter != snippet_.matches().end()) {
       if (match_iter->first > start) {
         // We're in a plain region.
